requirements.txt
Django==3.0.2
pyfortiapi==0.2.0
pycryptodome==3.9.4
django-auth-ldap==2.1.0
